Notification when user tapping instead of holding
