+++ react-web-app/src/assets/logo.jpg
