+++ react-web-app/src/components/TapNotification.jsx
@@ -1,0 +1,8 @@
+import React from 'react'
+export const TapNotification = ({ isTapping }) => {
+  return (
+    <div className={`tap-notification ${isTapping ? 'tap-notification--active' : ''}`}>
+      Please hold the button down.
+    </div>
+  )
+}
