+++ react-web-app/src/components/TapNotification.scss
@@ -1,0 +1,40 @@
+@import './Colors.scss';
+@import './Variables.scss';
+
+.tap-notification {
+  background-color: $accent-color;
+  color: white;
+  padding: 16px;
+  position: absolute;
+  bottom: calc(10vh + 40px);
+  font-size: 20px;
+  z-index: 2;
+  opacity: 0;
+  transform: scale(0);
+  border-radius: 10px;
+  -webkit-box-shadow: $light-shadow;
+  -moz-box-shadow: $light-shadow;
+  box-shadow: $light-shadow;
+  transform-origin: bottom center;
+  transition: all 0.5s;
+  transition-timing-function: ease-in-out;
+
+  &--active {
+    z-index: 2;
+    opacity: 1;
+    transform: scale(1);
+  }
+  &:after {
+    content: '';
+    position: absolute;
+    bottom: 0;
+    left: 50%;
+    width: 0;
+    height: 0;
+    border: 20px solid transparent;
+    border-top-color: $accent-color;
+    border-bottom: 0;
+    margin-left: -20px;
+    margin-bottom: -20px;
+  }
+}
