+++ react-web-app/src/components/Variables.scss
@@ -1,0 +1,2 @@
+$light-shadow: 3px 3px 16px 0px rgba(207, 207, 207, 1);
+$base-font-family: 'BrixSlabRegular';
